using same project structure as typescript-collections
