+++ .gitignore
@@ -2,12 +2,12 @@
 website/build/
 website/i18n/
 website/node_modules/
-website/package-lock.json
 
 # Codegen builds
-codegen/target/
-codegen/node_modules/
-codegen/package-lock.json
+npm-codegen/node_modules/
+npm-codegen/*/node_modules/
+npm-codegen/*/package-lock.json
+npm-codegen/dist/
 
 # Typedoc generated .md files
 docs/interfaces/
