--- codegen/.npmignore
@@ -1,7 +1,0 @@
-src
-target
-.prettierrc
-pom.template.xml
-tsconfig.json
-tslint.json
-
