--- codegen/tsconfig.json
@@ -1,17 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "es5",
-        "module": "commonjs",
-        "declaration": true,
-        "outDir": "./target/node",
-        "strict": true
-    },
-    "include": [
-        "../src/objects/*"
-    ],
-    "exclude": [
-        "node_modules",
-        "**/*.spec.ts",
-        "target"
-    ]
-}
