--- codegen/tslint.json
@@ -1,3 +1,0 @@
-{
-    "extends": ["tslint:recommended", "tslint-config-prettier"]
-}
