+++ mvn-codegen/package.json
@@ -1,22 +1,17 @@
 {
-  "name": "finos-fo-codegen",
+  "name": "finos-fo-codegen-mvn",
   "repository": "https://github.com/finos-fo/finos-fo.github.io.git",
   "license": "Apache-2.0",
   "files": ["target/node/*"],
   "scripts": {
-    "format": "prettier --write \"../src/objects/*.ts\" \"../src/objects/*.js\"",
-    "lint": "tslint -p tsconfig.json",
     "quicktype-java": "rm -rf target/java ; export DIR=$PWD/.. ; mkdir -p target/java/src/main/java/org/finos/fo ; cp pom.template.xml target/java/pom.xml ; cd target/java/src/main/java/org/finos/fo ; quicktype $DIR/src/objects/* --lang java --package org.finos.fo -o .",
     "quicktype-python": "rm -rf target/python ; export DIR=$PWD/.. ; mkdir -p target/python ; quicktype $DIR/src/objects/* --lang python -o target/python/financial-objects.py",
     "quicktype-cs": "rm -rf target/cs ; export DIR=$PWD/.. ; mkdir -p target/cs ; quicktype $DIR/src/objects/* --lang cs --namespace Finos.FinancialObjects -o target/cs/FinancialObjects.cs",
-    "build-java": "cd target/java ; mvn install",
-    "build-node": "rm -rf target/node ; tsc",
+    "java-bundle": "cd target/java ; mvn install",
     "print-java": "jar -tf target/java/target/*.jar",
     "deploy-java": "cd target/java ; mvn deploy"
   },
   "devDependencies": {
-    "quicktype": "15.0.171",
-    "tslint": "5.12.1",
-    "tslint-config-prettier": "1.17.0"
+    "quicktype": "15.0.171"
   }
 }
