+++ npm-codegen/.npmignore
@@ -1,0 +1,6 @@
+npm-sample
+local-sample
+tsconfig.json
+tslint.json
+yarn.lock
+README.md
