+++ npm-codegen/README.md
@@ -1,0 +1,24 @@
+# NPM code bundling
+
+The finanacial objects are packaged as NPM and published to npmjs.com
+
+## Building the package
+Run these commands from the `npm-codegen` folder:
+```
+yarn install
+yarn run node-bundle
+```
+
+## Running tests
+Run these commands from the `npm-codegen` folder:
+```
+cd samples
+yarn install
+yarn test
+```
+
+## Publish to NPM
+Run these commands from the `npm-codegen` folder:
+```
+npm init
+```
