+++ npm-codegen/local-sample/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "finos-fo-samples",
+  "repository": "https://github.com/finos-fo/finos-fo.github.io.git",
+  "license": "Apache-2.0",
+  "scripts": {
+    "install": "rm -rf node_modules ; mkdir -p node_modules/fo-test ; cp -Rf ../dist node_modules/fo-test/ ; cp ../package.json node_modules/fo-test",
+    "test": "ts-node printout.ts"
+  },
+  "devDependencies": {
+    "ts-node": "^7.0.1"
+  }
+}
