+++ src/objects/Equity.ts
@@ -1,4 +1,3 @@
 import { Product } from "./Product";
 
-export interface Equity extends Product {
-}
+export interface Equity extends Product {}
