+++ src/objects/IRS.ts
@@ -1,4 +1,3 @@
 import { Product } from "./Product";
 
-export interface IRS extends Product {
-}
+export interface IRS extends Product {}
