+++ src/objects/Product.ts
@@ -1,2 +1,1 @@
-export interface Product {
-}
+export interface Product {}
