+++ src/objects/RFQ.ts
@@ -6,10 +6,10 @@ import { Entity } from "./Entity";
  * Example: IRS, USD, 50m, 31/08/2018,31/08/2028, IMM, LCH, 2 way (Vanilla Swap), IRS, USD, 31/08/2018 10y/5y, IMM, LCH, 2 way (Vanilla Swap - 5y swap 10y fwd)
  */
 export interface RFQ extends Trade {
-    requestor: Entity;
-    direction: string;
-    notional: number;
-    startDate: Date;
-    endDate: Date;
-    rollConvention: string;  
+  requestor: Entity;
+  direction: string;
+  notional: number;
+  startDate: Date;
+  endDate: Date;
+  rollConvention: string;
 }
