+++ src/objects/RFQResponse.ts
@@ -4,7 +4,7 @@ import { RFQ } from "./RFQ";
  * Response to the RFQ object request
  */
 export interface RFQResponse extends RFQ {
-//   Responder Unique ID:[Entity B]
-// Receive Fixed Level: [x.xx]%
-// Pay Fixed Level: [x.xy]%
+  //   Responder Unique ID:[Entity B]
+  // Receive Fixed Level: [x.xx]%
+  // Pay Fixed Level: [x.xy]%
 }
